refactor: Rename classes and reorganize directories for budget input InputView -> BudgetInputView InputValidator -> BudgetValidator Move LottoConstants and BudgetValidator into subdirectories
