+++ src/main/java/lotto/utils/constant/LottoConstants.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.utils.constant;
 
 public class LottoConstants {
