+++ src/main/java/lotto/view/BudgetInputVIew.java
@@ -1,14 +1,14 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.utils.InputValidator;
+import lotto.utils.validator.BudgetValidator;
 
 public class BudgetInputVIew {
 
     public static int inputBudget() {
         System.out.println("구입 금액을 입력해주세요");
         String budget = Console.readLine();
-        InputValidator.validateInputBudget(budget);
+        BudgetValidator.validateInputBudget(budget);
         return Integer.parseInt(budget);
     }
